fix: improve download csv endpoint
