+++ pkg/handler/mine-service/mine-service.go
@@ -2,14 +2,12 @@ package mineservice
 
 import (
 	"net/http"
-	"os"
 	"strings"
 
 	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator/v10"
 	"github.com/workshopapps/pictureminer.api/internal/config"
 	"github.com/workshopapps/pictureminer.api/internal/model"
-	batchservice "github.com/workshopapps/pictureminer.api/service/batch-service"
 	mineservice "github.com/workshopapps/pictureminer.api/service/mine-service"
 	"github.com/workshopapps/pictureminer.api/utility"
 )
@@ -181,35 +179,6 @@ func (base *Controller) GetMinedImages(c *gin.Context) {
 
 func getFileName(url string) string {
 	return url[strings.LastIndex(url, "/")+1:]
-}
-
-func (base *Controller) DownloadCsv(c *gin.Context) {
-
-	// secretKey := config.GetConfig().Server.Secret
-	// token := utility.ExtractToken(c)
-	// userId, err := utility.GetKey("id", token, secretKey)
-	// if err != nil {
-	// 	rd := utility.BuildErrorResponse(http.StatusUnauthorized, "failed", "could not verify token", nil, gin.H{"error": err.Error()})
-	// 	c.JSON(http.StatusUnauthorized, rd)
-	// 	return
-	// }
-	batchId := c.Param("batchid")
-	var dummySlice, err = batchservice.GetImagesInBatch(batchId)
-	if err != nil {
-		rd := utility.BuildErrorResponse(http.StatusBadRequest, "failed", "could not get images for this batch id", nil, gin.H{"error": err.Error()})
-		c.JSON(http.StatusBadRequest, rd)
-		return
-	}
-	errr := mineservice.ParseImageResponseForDownload(dummySlice)
-	if errr != nil {
-		rd := utility.BuildErrorResponse(http.StatusBadRequest, "failed", "could not generate csv for download", nil, gin.H{"error": err.Error()})
-		c.JSON(http.StatusBadRequest, rd)
-		return
-	}
-
-	c.File("filename.csv")
-	defer os.Remove("filename.csv")
-
 }
 
 func (base *Controller) DeleteMinedImage(c *gin.Context) {
