+++ pkg/router/batch.go
@@ -19,6 +19,7 @@ func ProcessBatch(r *gin.Engine, validate *validator.Validate, ApiVersion string
 		authUrl.GET("/batch-service/get-batches", batch.GetBatches)
 		authUrl.POST("/batch-service/process-batch-api", batch.ProcessBatchAPI)
 		authUrl.GET("/batch-service/images/:batch_id", batch.GetBatchImages)
+		authUrl.GET("/batch-service/download/:batchid", batch.DownloadCsv)
 		authUrl.DELETE("/batch-service/delete/:id", batch.DeleteBatch)
 	}
