+++ service/batch-service/batch.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"os"
 	"path/filepath"
 	"strings"
 	"time"
@@ -327,4 +328,27 @@ func getUserEmail(userID string) (string, int, error) {
 		return "", http.StatusInternalServerError, err
 	}
 	return user.Email, http.StatusOK, nil
+}
+
+func ParseImageResponseForDownload(dt []model.BatchImage) error {
+	//create the file
+	file, err := os.Create("filename.csv")
+	if err != nil {
+		return err
+	}
+	writer := csv.NewWriter(file)
+
+	var line []string
+	l := append(line, "url")
+	l = append(l, "tag")
+	writer.Write(l)
+	//loop over the contents of the response
+	for _, value := range dt {
+
+		x := append(line, value.URL)
+		x = append(x, value.Tag)
+		writer.Write(x)
+	}
+	writer.Flush()
+	return nil
 }
